backend/requirements.txt
langchain-chroma
langchain-community
langchain-core
langchain-huggingface
langchain-openai
langchain-text-splitters
sentence-transformers
fastapi
uvicorn[standard]
python-dotenv
